job-finder-ai/requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.2
pandas==2.1.1
numpy==1.24.3
selenium==4.15.0
lxml==4.9.3
Pillow==10.0.1
jinja2==3.1.2
